==> Script_02_Analyzer.Rmd <==
---
title: "Data analysis for WP6"
author: "Patrizia Maier"
date: "`r format(Sys.time(), '%d %B, %Y')`"
output: pdf_document
---


```{r setup, include=F, cache=F}

library(tidyverse)
library(gtsummary)
library(rstatix)
library(kableExtra)
# install.packages('tinytex')
# tinytex::install_tinytex() # latex for pdf file creation

```


```{r load_data, include=F}

# load clinical, neuropsychological and scoring data (non-starmaze)
in_file <- "WP6_data/WP6_individual_data.Rdata"
load(in_file)
rm(in_file)

data_individual <- data_individual %>% 
  mutate(groupNo=factor(groupNo, levels=c(1,0), labels=c("ALS", "Ctrl")))

# load starmaze probe and clinical data 
in_file <- "WP6_data/WP6_starmaze_probe_data.Rdata"
load(in_file)
rm(in_file)

# subset starmaze successful probe data 
data_sm_suc <- data_sm %>% 
  filter(success==1)

```


# Sample demographics
Cross-sectional comparison between

* ALS (n = `r length(unique(data_individual$id[data_individual$groupNo=="1"]))`) 
* Control (n = `r length(unique(data_individual$id[data_individual$groupNo=="0"]))`)


```{r demo_all, echo=F, message=F, warning=F}

t1 <- data_individual %>% 
  select(groupNo, sex, ageYears, educationYears, languageGerman, SantaBarbaraSenseOfDirectionScale) %>%  
  tbl_summary(by=groupNo, 
              label=list(sex ~ "Sex", ageYears ~ "Age", educationYears ~ "Years total education",
                         languageGerman ~ "German native speaker", 
                         SantaBarbaraSenseOfDirectionScale ~ "Self-rated spatial abilities (SBSDS)"),
              statistic=list(all_continuous() ~ "{median} (IQR: {p25}-{p75})",
                             all_categorical() ~ "{n} ({p}%)"),
              digits=list(all_continuous() ~ c(1, 2, 1, 1)),
              missing="no") %>% 
  add_p(test=list(all_continuous() ~ "wilcox.test", all_categorical() ~  "chisq.test")) %>% 
  modify_header(label = "**Variable**")

t1 %>% as_flex_table()

# t1 %>%
#   as_flex_table() %>%
#   flextable::save_as_docx(path="WP6_data/ALS_CTRL_Demographics.docx")

```

```{r demo_clinical, echo=F, message=F, warning=F}

t2 <- data_individual %>% 
  filter(groupNo=="ALS") %>% 
  select(MN_involvement, ALS_variant, `ALSFRS-R`, `ALSFRS-R progression/month`, months_symptoms, months_diagnosis) %>% 
  tbl_summary(label=list(MN_involvement ~ "MN involvement", 
                         ALS_variant ~ "ALS (bulbar/spinal)", 
                         months_symptoms ~ "Months since inital symptoms",
                         months_diagnosis ~ "Months since diagnosis"),
              statistic=list(all_continuous() ~ "{median} (IQR: {p25}-{p75})",
                             all_categorical() ~ "{n} ({p}%)"),
              digits=list(all_continuous() ~ c(1, 1))) %>% 
  add_n() %>%
  modify_header(label = "**Variable**")

t2 %>% as_flex_table()

# t2 %>%
#   as_flex_table() %>%
#   flextable::save_as_docx(path="WP6_data/ALS_Clinical.docx")

```

# Results 

## Post-navigational scoring 

```{r results_scoring, echo=F, message=F, warning=F}

effect_size <- function(data, variable, by, ...) {
  rstatix::wilcox_effsize(data, as.formula(glue::glue("{variable} ~ {by}")))$effsize
}

t3 <- data_individual %>% 
  select(groupNo, POST_mazeReconstruction, POST_objectIdentity, POST_objectLocation) %>% 
  tbl_summary(by=groupNo,
              type=list(everything() ~ 'continuous'),
              label=list(POST_objectIdentity ~ "Object identity", 
                         POST_objectLocation ~ "Object location", 
                         POST_mazeReconstruction ~ "Maze reconstruction"),
              statistic=list(all_continuous() ~ "{median} ({p25}-{p75})"),
              digits=list(all_continuous() ~ c(2, 2)),
              missing="no") %>% 
  add_stat(fns=everything() ~ effect_size) %>%
  add_p(test=list(everything() ~ "wilcox.test"), 
        pvalue_fun=function(x) style_number(x, digits=3)) %>%  
  modify_header(list(label="**Post-navigation memory**", add_stat_1="**effect size**"))

t3 %>% as_flex_table()

# t3 %>%
#   as_flex_table() %>%
#   flextable::save_as_docx(path="WP6_data/ALS_CTRL_Scoring.docx")

```

## Neuropsychological data 

```{r results_np, echo=F, message=F, warning=F}

effect_size <- function(data, variable, by, ...) {
  rstatix::wilcox_effsize(data, as.formula(glue::glue("{variable} ~ {by}")))$effsize
}

t4 <- data_individual %>% 
  select(groupNo, ECAS_executive, ECAS_verbalFluency, ECAS_language, ECAS_verbalMemory, ECAS_visuospatial, 
         SPART_immediateMemory, SPART_delayedMemory, SPART_overallMemory, 
         `5PT`, PTSOT) %>% 
  tbl_summary(by=groupNo,
              type=list(everything() ~ 'continuous'),
              label=list(ECAS_executive ~ "ECAS executive",
                         ECAS_verbalFluency ~ "ECAS fluency",
                         ECAS_language ~ "ECAS language",
                         ECAS_verbalMemory ~ "ECAS memory", 
                         ECAS_visuospatial ~ "ECAS visuospatial", 
                         SPART_immediateMemory ~ "SPART immediate",
                         SPART_delayedMemory ~ "SPART delayed",
                         SPART_overallMemory ~ "SPART total memory",
                         `5PT` ~ "5PT fluency",
                         PTSOT ~ "PTSOT orientation"),
              statistic=list(everything() ~ "{median} ({p25}-{p75})"),
              digits=list(everything() ~ c(2, 2)),
              missing="no") %>% 
  add_stat(fns=everything() ~ effect_size) %>%
  add_p(test=list(everything() ~ "wilcox.test"), 
        pvalue_fun=function(x) style_number(x, digits=3)) %>% 
  modify_header(list(label="**Neuropsychology**", add_stat_1="**effect size**"))

t4 %>% as_flex_table()

# t4 %>%
#   as_flex_table() %>%
#   flextable::save_as_docx(path="WP6_data/ALS_CTRL_Neuropsychology.docx")

```

```{r results_np_impairment, echo=F, message=F, warning=F}

ecas_impaired <- data_individual %>% 
  select(id, groupNo, ends_with("below_cut"))

t5 <- ecas_impaired %>% 
  select(-id) %>% 
  tbl_summary(by=groupNo) %>% 
  add_n()

t5 %>% as_flex_table()

ecas_impaired <- ecas_impaired %>% 
  drop_na()

```

Participants below executive cut-off: `r ecas_impaired$id[ecas_impaired$ECAS_sub_executive_below_cut]`

Participants below fluency cut-off: `r ecas_impaired$id[ecas_impaired$ECAS_sub_verbal_fluency_below_cut]`

Participants below language cut-off: `r ecas_impaired$id[ecas_impaired$ECAS_sub_language_below_cut]`

Participants below memory cut-off: `r ecas_impaired$id[ecas_impaired$ECAS_sub_memory_below_cut]`

Participants below visuospatial cut-off: `r ecas_impaired$id[ecas_impaired$ECAS_sub_spatial_below_cut]`


## Correlations

```{r results_correlation, echo=F, message=F, warning=F}

data_corr <- data_individual  %>% select(id, groupNo, ECAS_total_score, ECAS_sub_memory, 
                                    dfb_q1_sex, dfb_q2_age, dfb_q3_years_edu_total,
                                    `ALS-FRS-R`, `FRS-/Monat`, is_t1_months, id_t1_months) %>%
  drop_na(ECAS_sub_memory) %>% filter(groupNo=="ALS")

ggplot(data_corr, aes(x=`ALS-FRS-R`, y=ECAS_total_score)) + geom_smooth(method="lm") + geom_point()
cor.test(data_corr$ECAS_total_score, data_corr$`ALS-FRS-R`, method="spearman")

ggplot(data_corr, aes(x=`FRS-/Monat`, y=ECAS_total_score)) + geom_smooth(method="lm") + geom_point()
cor.test(data_corr$ECAS_total_score, data_corr$`FRS-/Monat`, method="spearman")

ggplot(data_corr, aes(x=`ALS-FRS-R`, y=ECAS_sub_memory)) + geom_smooth(method="lm") + geom_point()
cor.test(data_corr$ECAS_sub_memory, data_corr$`ALS-FRS-R`, method="spearman")

ggplot(data_corr, aes(x=`FRS-/Monat`, y=ECAS_sub_memory)) + geom_smooth(method="lm") + geom_point()
cor.test(data_corr$ECAS_sub_memory, data_corr$`FRS-/Monat`, method="spearman")

ggplot(data_corr, aes(x=is_t1_months, y=ECAS_sub_memory)) + geom_smooth(method="lm") + geom_point()
cor.test(data_corr$ECAS_sub_memory, data_corr$is_t1_months, method="spearman")

ggplot(data_corr, aes(x=id_t1_months, y=ECAS_sub_memory)) + geom_smooth(method="lm") + geom_point()
cor.test(data_corr$ECAS_sub_memory, data_corr$id_t1_months, method="spearman")

ggplot(data_corr, aes(x=dfb_q2_age, y=ECAS_sub_memory)) + geom_smooth(method="lm") + geom_point()
cor.test(data_corr$ECAS_sub_memory, data_corr$dfb_q2_age, method="spearman")

ggplot(data_corr, aes(x=dfb_q3_years_edu_total, y=ECAS_sub_memory)) + geom_smooth(method="lm") + geom_point()
cor.test(data_corr$ECAS_sub_memory, data_corr$dfb_q3_years_edu_total, method="spearman")

m0 <- glm(ECAS_sub_memory ~ dfb_q1_sex + dfb_q2_age + dfb_q3_years_edu_total + 
            `ALS-FRS-R` + `FRS-/Monat` + is_t1_months, data=data_corr) 
summary(m0)
performance::r2(m0)

rm(m0, data_corr)
```

# Supplementary analyses 

## ALS, PLS vs. PMA

```{r results_supplement_MN, echo=F, message=F, warning=F}

# ALS, PLS vs. PMA 
data_mn <- data_sm %>% 
  group_by(id, groupNo, trialCondition, MN_involvement) %>% 
  summarize_at(vars(success), mean, na.rm=T) %>% 
  ungroup()

data_mn_suc <- data_sm_suc %>% 
  group_by(id, groupNo, trialCondition, MN_involvement) %>% 
  summarize_at(vars(latency_seconds, pathError_percent, searchAccuracy_percent), mean, na.rm=T) %>% 
  ungroup()

# condition-wise
data_mn_joint <- data_mn %>% 
  left_join(data_mn_suc)
rm(data_mn, data_mn_suc)

ts5 <- data_mn_joint %>% 
  select(MN_involvement, trialCondition, success, latency_seconds, pathError_percent, searchAccuracy_percent) %>%  
  tbl_strata(strata=trialCondition,
             .tbl_fun =
               ~ .x %>% tbl_summary(by=MN_involvement,
                                    type=list(everything() ~ 'continuous'),
                                    label=list(success ~ "Success rate",
                                               latency_seconds ~ "Latency",
                                               pathError_percent ~ "Path error",
                                               searchAccuracy_percent ~ "Search accuracy"), 
                                    statistic=list(everything() ~ "{median} ({p25}-{p75})"),
                                    digits=list(everything() ~ c(2, 1, 1)),
                                    missing="no") %>% 
               add_p(test=list(everything() ~ "kruskal.test"), 
                     pvalue_fun=function(x) style_number(x, digits=3)),
             .combine_with = "tbl_stack")
             

# ts5 %>%
#   as_flex_table() %>%
#   flextable::save_as_docx(path="WP6_data/Supplement_ALS_PLS_PMA.docx")
rm(ts5)

# averaged 
data_mn_joint_avg <- data_mn_joint %>% 
  group_by(id, groupNo, MN_involvement) %>% 
  summarize_at(vars(success, latency_seconds, pathError_percent, searchAccuracy_percent), mean, na.rm=T) %>% 
  ungroup()

ts5 <- data_mn_joint_avg %>% 
  select(MN_involvement, success, latency_seconds, pathError_percent, searchAccuracy_percent) %>%  
  tbl_summary(by=MN_involvement, 
              type=list(everything() ~ 'continuous'),
              label=list(success ~ "Success rate",
                         latency_seconds ~ "Latency",
                         pathError_percent ~ "Path error",
                         searchAccuracy_percent ~ "Search accuracy"), 
              statistic=list(everything() ~ "{median} ({p25}-{p75})"),
              digits=list(everything() ~ c(2, 1, 1)),
              missing="no") %>%
  add_p(test=list(everything() ~ "kruskal.test"), 
        pvalue_fun=function(x) style_number(x, digits=3))

# ts5 %>%
#   as_flex_table() %>%
#   flextable::save_as_docx(path="WP6_data/Supplement_ALS_PLS_PMA_avg.docx")
rm(ts5)
```

## ALSci vs. non-ci

```{r results_supplement_ALSci, echo=F, message=F, warning=F}

# ALSci vs. ALS normal 
data_ci <- data_sm %>% 
  group_by(id, groupNo, trialCondition, ALSci) %>% 
  summarize_at(vars(success), mean, na.rm=T) %>% 
  ungroup()

data_ci_suc <- data_sm_suc %>% 
  group_by(id, groupNo, trialCondition, ALSci) %>% 
  summarize_at(vars(latency_seconds, pathError_percent, searchAccuracy_percent), mean, na.rm=T) %>% 
  ungroup()

# condition-wise
data_ci_joint <- data_ci %>% 
  left_join(data_ci_suc)
rm(data_ci, data_ci_suc)

ts4 <- data_ci_joint %>% 
  select(ALSci, trialCondition, success, latency_seconds, pathError_percent, searchAccuracy_percent) %>%  
  tbl_strata(strata=trialCondition,
             .tbl_fun =
               ~ .x %>% tbl_summary(by=ALSci,
                                    type=list(everything() ~ 'continuous'),
                                    label=list(success ~ "Success rate",
                                               latency_seconds ~ "Latency",
                                               pathError_percent ~ "Path error",
                                               searchAccuracy_percent ~ "Search accuracy"), 
                                    statistic=list(everything() ~ "{median} ({p25}-{p75})"),
                                    digits=list(everything() ~ c(2, 1, 1)),
                                    missing="no") %>% 
               add_p(test=list(everything() ~ "kruskal.test"), 
                     pvalue_fun=function(x) style_number(x, digits=3)),
             .combine_with = "tbl_stack")
             

# ts4 %>%
#   as_flex_table() %>%
#   flextable::save_as_docx(path="WP6_data/Supplement_ALSci.docx")
rm(ts4)

# averaged 
data_ci_joint_avg <- data_ci_joint %>% 
  group_by(id, groupNo, ALSci) %>% 
  summarize_at(vars(success, latency_seconds, pathError_percent, searchAccuracy_percent), mean, na.rm=T) %>% 
  ungroup()

ts4 <- data_ci_joint_avg %>% 
  select(ALSci, success, latency_seconds, pathError_percent, searchAccuracy_percent) %>%  
  tbl_summary(by=ALSci, 
              type=list(everything() ~ 'continuous'),
              label=list(success ~ "Success rate",
                         latency_seconds ~ "Latency",
                         pathError_percent ~ "Path error",
                         searchAccuracy_percent ~ "Search accuracy"), 
              statistic=list(everything() ~ "{median} ({p25}-{p75})"),
              digits=list(everything() ~ c(2, 1, 1)),
              missing="no") %>%
  add_p(test=list(everything() ~ "kruskal.test"), 
        pvalue_fun=function(x) style_number(x, digits=3))

# ts4 %>%
#   as_flex_table() %>%
#   flextable::save_as_docx(path="WP6_data/Supplement_ALSci_avg.docx")
rm(ts4, data_ci_joint, data_ci_joint_avg)
```

# General information on recruting success

```{r recruiting, echo=F, message=F, warning=F}

read_sheets <- function(filename, tibble = FALSE) {
  sheets <- readxl::excel_sheets(filename)
  x <- lapply(sheets, function(X) readxl::read_excel(filename, sheet = X, skip=2))
  if(!tibble) x <- lapply(x, as.data.frame)
  names(x) <- sheets
  x
}

filename1 <- "WP6_data/Rekrutierung_210820.xlsx"
mysheets <- read_sheets(filename1)
names_invite <- do.call(rbind.data.frame, mysheets) %>% 
  select(-c(Alter, Adresse, `Termin Tag`, `Termin Uhrzeit`, Rückmeldung, Notizen)) %>% 
  filter(`Brief verschickt` != "nein") %>% 
  select(-c(`Brief verschickt`)) %>% 
  unique() %>% 
  unite(Vorname, Nachname, col="Name", sep=" ") %>% 
  arrange(Name) 

filename2 <- "WP6_data/WP6_id_List_6200_6399.xlsx"
names_participate <- readxl::read_excel(filename2, skip=1, n_max=52, trim_ws = T) %>% 
  select(Name) %>% 
  arrange(Name)

rm(filename1, filename2, read_sheets)
```

## People that were recruited via mail (with help of ALS-Ambulanz)
N of letters = `r length(names_invite$Name) `

N of respondents = `r length(names_invite$Name[names_invite$Name %in% names_participate$Name]) `
(`r round(length(names_invite$Name[names_invite$Name %in% names_participate$Name])/length(names_invite$Name)*100, digits=2)`)

## People that were recruited elsewhere (e.g. from the Neurology ward)
N of respondents = `r length(names_participate$Name[!names_participate$Name %in% names_invite$Name]) `
